+++ Packages/Core/Sources/API/Common/Query/Query+Option.swift
@@ -40,42 +40,3 @@ public extension Query {
         }
     }
 }
-
-public extension Query.Option {
-    struct Pagination: Sendable, Equatable, Hashable {
-        let page: Int
-        let pageSize: Int
-
-        func encode(to container: inout KeyedEncodingContainer<DynamicCodingKey>) throws {
-            try container.encode(page, forKey: "page")
-            try container.encode(pageSize, forKey: "limit")
-        }
-    }
-}
-
-public extension Query.Option {
-    struct Sort: Sendable, Equatable, Hashable, Encodable {
-        public let field: Item.Field
-        public let order: SortOrder
-
-        private init(field: Item.Field, order: SortOrder) {
-            self.field = field
-            self.order = order
-        }
-
-        public static func by(_ field: Item.Field, _ order: SortOrder = .forward) -> Sort {
-            .init(field: field, order: order)
-        }
-
-        public func encode(to encoder: any Encoder) throws {
-            var container = encoder.unkeyedContainer()
-            try container.encode(field.rawValue)
-            switch order {
-            case .forward:
-                try container.encode("asc")
-            case .reverse:
-                try container.encode("desc")
-            }
-        }
-    }
-}
